python-fastapi-s3aws-api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
qrcode==8.0
boto3==1.37.23
pillow==11.1.0